server/requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.3
uvicorn==0.38.0
Pillow>=9.0.0